dnsdist: Move DownstreamConnectionsManager to its own header
